added bonus functions

+++ libft/ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: knieve-l <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/16 12:36:50 by knieve-l          #+#    #+#             */
-/*   Updated: 2024/09/20 11:18:11 by knieve-l         ###   ########.fr       */
+/*   Updated: 2024/09/23 20:25:51 by kamil            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int i;
+	unsigned int	i;
 
 	i = 0;
 	while (i < n && (s1[i] != '\0' || s2[i] != '\0'))
